FinSum AI-Powered Financial Report Summarizer/financial-summarizer/requirements.txt
requests==2.31.0
pandas==2.0.3
beautifulsoup4==4.12.2
nltk==3.8.1
transformers==4.35.0
torch==2.0.1
flask==2.3.3
werkzeug==2.3.7
flask-restx==1.1.0
marshmallow==3.20.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyyaml==6.0.1
streamlit==1.32.0
gunicorn==21.2.0
prometheus-client==0.19.0
rouge-score==0.1.2
pytest==7.4.3
logging-formatter-anticrlf==1.2 